requirements.txt
flask-login
flask-SQLAlchemy
flask-admin